limits the evaluation of while loops
For `x86/x86-64` arch `while` loops weren't used,
and loops were implemented as a jump to the beginning
of an instruction.

Now it's not true, since it's guaranteed that `while` loops can
have only data effects. And instructions with `REP` prefixes
are implemented with using of `while` loop without any jumps.

But that doesn't work with our testsuite framework, where
every loop iteration is represented as an independent instruction
with its own set of events, which we verified previosly after each
iteration. And with new inplementation we can get stuck with
nearly infinite loop.

That's why we'll evaluate only first iteration of each while loop.
in order to restore the previous behavior.